backend/requirements.txt
flask
requests
pyserial
geocoder
python-dotenv
